AC: split the DualCool TLV handling into an ac_common base class
RAC_056905_WW is not one model's handler. It is the DualCool TLV scheme - 0x1f7
power, 0x1f9 mode, 0x1fa fan, 0x1fd/0x1fe temperatures, the 0x2cc / 0x2cd / 0x2d3
capability bitmaps, the pipe and ODU diagnostics, the basic-filter priv-command -
with one model's particulars mixed into it. Anything else on the same scheme could
either copy the file or be translated into RAC's expectations, and both have been
tried in this project.

So the protocol moves to ac_common and RAC keeps what is actually its own: its
0x1f9 / 0x1fa scales, its setpoint range, the +50 bias on its power reading, and the
fact that it carries a basic-filter counter. Nothing here is tuned to a unit; where
models are known to differ, the difference is stated by the subclass as an override.

Behaviour-preserving, including the two fixes RAC took last week, which move across
rather than being lost in the shuffle:

  - the filter reset queries the counter before clearing it (02cf15c), so a reset
    does not record a usage figure up to 24 hours short of what the filter ran
  - a hook already on modeChangeHooks is not also registered on powerChangeHooks
    (2641ccd), since power reads as a mode change too and the appliance was getting
    the same write twice on every power-up

Both are now covered by RAC tests that fail without them: one counts the jet writes
a power-up produces, the other reads the priv-command subcommands and asserts the
query goes out before the reset.
